+++ packages/chameleon-loader/src/loader.js
@@ -49,8 +49,26 @@ module.exports = function (content) {
   }
 
   //loader的类型  wx  web weex
-  const {cmlType, media, builtinNpmName, cmss = defaultCmss, isInjectBaseStyle = true,isWrapComponent = true} = options;
-  options.isInjectBaseStyle = isInjectBaseStyle;
+  const {cmlType, media, builtinNpmName, cmss = defaultCmss, isWrapComponent = true, subProject = []} = options;
+  let { isInjectBaseStyle = true } = options;
+  //处理拿到json对象, 使用baseStyle来配置是否注入基础样式
+  jsonObject = cmlUtils.getJsonFileContent(self.resourcePath, cmlType);
+  
+  // 处理子项目的isInjectBaseStyle
+  if (subProject.length) {
+    subProject.forEach(item => {
+      if (self.resourcePath.indexOf(item.npmName) > -1 && item.isInjectBaseStyle !== undefined) {
+        isInjectBaseStyle = item.isInjectBaseStyle;
+      }
+    })
+  }
+  
+
+  if (jsonObject && jsonObject.baseStyle !== undefined) {
+    isInjectBaseStyle = jsonObject.baseStyle;
+  }
+
+  
   if(isInjectBaseStyle && cmlType === 'weex') {
     content = prehandle.injectWeexBaseStyle(content, self);
   }
@@ -74,8 +92,6 @@ module.exports = function (content) {
     process.cwd()
   )
 
-  //处理拿到json对象
-  jsonObject = cmlUtils.getJsonFileContent(self.resourcePath, cmlType);
   //是否是引用的原生小程序组件  wxml文件
   const isWxmlComponent = extName === '.wxml';
   const isAxmlComponent = extName === '.axml';
@@ -145,10 +161,11 @@ module.exports = function (content) {
     wx: 'wxml',
     alipay: 'axml',
     baidu: 'swan',
+    qq: 'qml'
   }
   //小程序模板后缀正则
   const miniTplExtReg = /(\.wxml|\.axml)$/;
-  const miniCmlReg = /(\.cml|\.wx\.cml|\.alipay\.cml| |\.baidu\.cml)$/;
+  const miniCmlReg = /(\.cml|\.wx\.cml|\.alipay\.cml|\.qq\.cml|\.baidu\.cml)$/;
 
   if(isMiniAppRawComponent) {
     miniAppRawComponentHandler.call(this);
@@ -156,6 +173,7 @@ module.exports = function (content) {
       //handler中改变output的值 最后返回output
       switch (cmlType) {
         case 'wx':
+        case 'qq':
         case 'alipay':
         case 'baidu':
           miniAppHandler.call(this);
@@ -191,7 +209,7 @@ module.exports = function (content) {
   // 引用微信小程序组件处理
   function miniAppRawComponentHandler() {
     
-    if((cmlType === 'wx' && extName === '.wxml') || (cmlType === 'alipay' && extName === '.axml') || (cmlType === 'baidu' && extName === '.swan')) {
+    if((cmlType === 'wx' && extName === '.wxml') || (cmlType === 'alipay' && extName === '.axml') || (cmlType === 'baidu' && extName === '.swan') || (cmlType === 'qq' && extName === '.qml')) {
       //生成json文件
       let jsonFile = filePath.replace(miniTplExtReg,'.json');
       if(!cmlUtils.isFile(jsonFile)) {