+++ packages/chameleon-tool-utils/src/index.js
@@ -296,7 +296,7 @@ _.getJsonFileContent = function (filePath, confType) {
       if (copyNpm && copyNpm.length > 0) {
         copyNpm.forEach(function(npmName) {
           let packageJson = JSON.parse(fs.readFileSync(path.join(cml.projectRoot, 'node_modules', npmName, 'package.json'), {encoding: 'utf-8'}));
-          let cmlConfig = packageJson.cml && packageJson.cml[confType]; 
+          let cmlConfig = packageJson.cml && packageJson.cml[confType];
           if (cmlConfig && cmlConfig.pages && cmlConfig.pages.length > 0) {
             cmlConfig.pages.forEach(item => {
               if (!~targetObject.pages.indexOf(item)) {
@@ -311,7 +311,9 @@ _.getJsonFileContent = function (filePath, confType) {
       // 处理subProject配置的npm包中cml项目的页面
       let subProject = cml.config.get().subProject;
       if (subProject && subProject.length > 0) {
-        subProject.forEach(function(npmName) {
+        subProject.forEach(function(item) {
+          
+          let { npmName } = item;
           let npmRouterConfig = _.readsubProjectRouterConfig(cml.projectRoot, npmName);
           npmRouterConfig.routes && npmRouterConfig.routes.forEach(item => {
             let cmlFilePath = path.join(cml.projectRoot, 'node_modules', npmName, 'src', item.path + '.cml');
@@ -586,7 +588,7 @@ _.getOnePackageComponents = function (npmName, packageFilePath, cmlType, context
     let globPath = path.join(context, 'node_modules', npmName, main, '/**/*.cml');
 
     // 需要忽略掉的组件
-    let ignoreComponents = ['web', 'weex', 'wx', 'alipay', 'baidu'];
+    let ignoreComponents = ['web', 'weex', 'wx', 'alipay', 'baidu', 'qq'];
     if (~ignoreComponents.indexOf(cmlType)) {
       ignoreComponents.splice(ignoreComponents.indexOf(cmlType), 1);
     }
@@ -700,6 +702,10 @@ _.handleComponentUrl = function (context, cmlFilePath, comPath, cmlType) {
     refUrl = refUrl.replace(/\.swan$/g, '');
   }
 
+  if (cmlType === 'qq') {
+    refUrl = refUrl.replace(/\.qml$/g, '');
+  }
+
   return {
     refUrl,
     filePath
@@ -733,7 +739,8 @@ _.findComponent = function (filePath, cmlType) {
     web: ['.vue', '.js'],
     wx: '.wxml',
     baidu: '.swan',
-    alipay: '.axml'
+    alipay: '.axml',
+    qq: 'qml'
   }
 
   let ext = fileExtMap[cmlType];
@@ -751,7 +758,7 @@ _.findComponent = function (filePath, cmlType) {
 
 // 提供给cml-lint使用 cml-lint不知道cmlType
 _.lintHandleComponentUrl = function(context, cmlFilePath, comPath) {
-  let cmlTypeList = ['wx', 'web', 'weex', 'alipay', 'baidu'];
+  let cmlTypeList = ['wx', 'web', 'weex', 'alipay', 'baidu', 'qq'];
   for (let i = 0; i < cmlTypeList.length; i++) {
     let cmlType = cmlTypeList[i];
     let result = _.handleComponentUrl(context, cmlFilePath, comPath, cmlType);
@@ -808,7 +815,7 @@ _.npmComponentRefPath = function (componentAbsolutePath, context) {
   if (refUrl[0] !== '/') {
     refUrl = '/' + refUrl
   }
-  refUrl = refUrl.replace(/(\.cml|\.web\.cml|\.alipay\.cml|\.baidu\.cml|\.wx\.cml|\.weex\.cml)/, '');
+  refUrl = refUrl.replace(/(\.cml|\.web\.cml|\.alipay\.cml|\.baidu\.cml|\.wx\.cml|\.weex\.cml|\.qq\.cml)/, '');
   return refUrl;
 
 }
@@ -907,7 +914,7 @@ _.getExportEntry = function (cmlType, context, entry = []) {
       } else if (_.isDirectory(filePath)) {
         filePath = path.join(filePath, '**/*.cml');
         // 需要忽略掉的组件
-        let ignoreComponents = ['web', 'weex', 'wx', 'alipay', 'baidu'];
+        let ignoreComponents = ['web', 'weex', 'wx', 'alipay', 'baidu', 'qq'];
         if (~ignoreComponents.indexOf(cmlType)) {
           ignoreComponents.splice(ignoreComponents.indexOf(cmlType), 1);
         }
@@ -950,6 +957,9 @@ _.getPureEntryName = function (cmlFilePath, cmlType, context) {
   if (cmlType === 'baidu') {
     entryPath = entryPath.replace(/\.swan/g, '');
   }
+  if (cmlType === 'qq') {
+    entryPath = entryPath.replace(/\.qml/g, '');
+  }
   return entryPath.replace(cmlExtReg, '');
 }
 
@@ -989,7 +999,7 @@ _.getCmlFileType = function(cmlFilePath, context, cmlType) {
       type = 'app';
     } else {
       let subProject = cml.config.get().subProject || [];
-      let npmNames = subProject.map(item => 'node_modules/' + item);
+      let npmNames = subProject.map(item => 'node_modules/' + item.npmName);
       let subProjectIndex = -1;
       for (let i = 0; i < npmNames.length; i++) {
         if (~cmlFilePath.indexOf(npmNames[i])) {
@@ -999,10 +1009,8 @@ _.getCmlFileType = function(cmlFilePath, context, cmlType) {
       }
       // 是subProject npm包中的cml文件 用subProject中的router.config.json判断
       if (subProjectIndex != -1) {
-        let routerConfig = _.readsubProjectRouterConfig(context, subProject[subProjectIndex]);
-        let pageFiles = routerConfig.routes.map(item => {
-          return path.join(context, 'node_modules', subProject[subProjectIndex], 'src', item.path + '.cml');
-        })
+        let routerConfig = _.readsubProjectRouterConfig(context, subProject[subProjectIndex].npmName);
+        let pageFiles = routerConfig.routes.map(item => path.join(context, 'node_modules', subProject[subProjectIndex].npmName, 'src', item.path + '.cml'))
         // 如果是配置的路由则是page
         if (~pageFiles.indexOf(cmlFilePath)) {
           type = 'page';