+++ packages/chameleon-tool/configs/component_export/getWebExportConfig_new.js
@@ -14,7 +14,8 @@ module.exports = function(options) {
     media,
     root,
     disableExtract,
-    mode = 'production'
+    mode = 'production',
+    externals
   } = options;
 
   function getstaticPath(filetype) {
@@ -175,6 +176,11 @@ module.exports = function(options) {
       })
     )
   }
+
+  if (externals) {
+    exportConfig.externals = externals;
+  }
+
 
   return merge.smart(getWebCommonConfig(options), exportConfig);