+++ CHANGELOG.md
@@ -1,13 +1,41 @@
+## [0.3.3-alpha.6]
+### Features
+  - 组件导出依赖分离
+  - 可通过命令行配置 preview 预览页是否打开
+  
+## [0.3.3-alpha.5]
+### Features
+  - 合并qq小程序分支
+  - web端原生origin click事件不处理成tap
+  
 ## [0.3.3-alpha.4]
+### Features
   - 修改全局变量校验失效问题，同时增加 qq 全局变量名单。
   - 模板变量校验添加 LogicalExpression 处理逻辑表达式里的变量
+### Bug Fixes
   - 修改cover-view报未定义组件问题件
   - 修复引入zepto库后tap事件触发两次问题
+  
 ## [0.3.3-alpha.3]
+### Features
   - 支持web weex可配置是否包裹组件
+
+## [0.3.3-alpha.qq.1]
+### Features
+* 支持qq小程序
+  旧项目支持qq小程序 需要修改的地方
+  - 1 升级以下几个npm包版本
+    "chameleon-api": "0.5.0-alpha.6",
+    "chameleon-bridge": "0.2.0-alpha.5",
+    "chameleon-runtime": "0.2.2-alpha.qq",
+    "chameleon-store": "0.1.0-alpha.qq",
+    "chameleon-ui-builtin": "0.2.11-alpha.qq",
+    "cml-ui": "0.2.0-alpha.qq"
+  - 2 chameleon.config.js中`platforms`字段添加`qq`
+  - 
 ## [0.3.3-alpha.2]
 
-###  Bug Fixes 
+### Bug Fixes 
 
 * 修复 weex dev模式liveload失效
   老项目如果修复，还需要升级项目中两个npm包如下：